Implement Single Responsability
