+++ app/routes/home.py
@@ -1,0 +1,5 @@
+from app import app
+
+@app.route('/')
+def index():
+    return 'Welcome to the Catalog System!'
